REFACTOR: NAV-87 - Remove model package for raptor.
