+++ src/main/java/ch/naviqore/raptor/RaptorBuilder.java
@@ -1,4 +1,4 @@
-package ch.naviqore.raptor.model;
+package ch.naviqore.raptor;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
