+++ src/main/java/ch/naviqore/raptor/RouteStop.java
@@ -1,4 +1,4 @@
-package ch.naviqore.raptor.model;
+package ch.naviqore.raptor;
 
 record RouteStop(int stopIndex, int routeIndex) {
 }
