+++ src/main/java/ch/naviqore/raptor/StopContext.java
@@ -1,4 +1,4 @@
-package ch.naviqore.raptor.model;
+package ch.naviqore.raptor;
 
 record StopContext(Transfer[] transfers, Stop[] stops, int[] stopRoutes) {
 }
