+++ src/main/java/ch/naviqore/raptor/Transfer.java
@@ -1,4 +1,4 @@
-package ch.naviqore.raptor.model;
+package ch.naviqore.raptor;
 
 record Transfer(int targetStopIdx, int duration) {
 }
