+++ src/test/java/ch/naviqore/Benchmark.java
@@ -7,8 +7,8 @@ import ch.naviqore.gtfs.schedule.model.Stop;
 import ch.naviqore.gtfs.schedule.model.StopTime;
 import ch.naviqore.gtfs.schedule.model.Trip;
 import ch.naviqore.gtfs.schedule.type.ServiceDayTime;
-import ch.naviqore.raptor.model.Connection;
-import ch.naviqore.raptor.model.Raptor;
+import ch.naviqore.raptor.Connection;
+import ch.naviqore.raptor.Raptor;
 import ch.naviqore.service.impl.GtfsToRaptorConverter;
 import ch.naviqore.service.impl.transfergenerator.MinimumTimeTransfer;
 import ch.naviqore.service.impl.transfergenerator.SameStationTransferGenerator;
