+++ src/test/java/ch/naviqore/raptor/model/RaptorTestBuilder.java
@@ -1,5 +1,7 @@
 package ch.naviqore.raptor.model;
 
+import ch.naviqore.raptor.Raptor;
+import ch.naviqore.raptor.RaptorBuilder;
 import lombok.NoArgsConstructor;
 
 import java.util.ArrayList;
