+++ src/test/java/ch/naviqore/raptor/model/RouteBuilderTest.java
@@ -1,5 +1,7 @@
 package ch.naviqore.raptor.model;
 
+import ch.naviqore.raptor.RouteBuilder;
+import ch.naviqore.raptor.StopTime;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
