+++ src/test/java/ch/naviqore/service/impl/GtfsToRaptorConverterIT.java
@@ -3,7 +3,7 @@ package ch.naviqore.service.impl;
 import ch.naviqore.gtfs.schedule.GtfsScheduleReader;
 import ch.naviqore.gtfs.schedule.GtfsScheduleTestData;
 import ch.naviqore.gtfs.schedule.model.GtfsSchedule;
-import ch.naviqore.raptor.model.Raptor;
+import ch.naviqore.raptor.Raptor;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
